Move singleflight out of middleware/
It is not middleware, so move it up one level. It was also use from
core/ showing that indeed its use is not limited to middlewares.
