+++ core/setup/etcd.go
@@ -11,7 +11,7 @@ import (
 	"github.com/miekg/coredns/middleware"
 	"github.com/miekg/coredns/middleware/etcd"
 	"github.com/miekg/coredns/middleware/proxy"
-	"github.com/miekg/coredns/middleware/singleflight"
+	"github.com/miekg/coredns/singleflight"
 
 	etcdc "github.com/coreos/etcd/client"
 	"golang.org/x/net/context"
