+++ middleware/etcd/etcd.go
@@ -10,7 +10,7 @@ import (
 	"github.com/miekg/coredns/middleware"
 	"github.com/miekg/coredns/middleware/etcd/msg"
 	"github.com/miekg/coredns/middleware/proxy"
-	"github.com/miekg/coredns/middleware/singleflight"
+	"github.com/miekg/coredns/singleflight"
 
 	etcdc "github.com/coreos/etcd/client"
 	"golang.org/x/net/context"
